# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Это API (swagger 2.0) системы автоматизации психолого-педагогическ\
    ого сопровождения субъектов образовательной деятельности для школы №18."
  version: "1.0.7"
  title: "PsychoSys18"
  contact:
    email: "antony.nor@yandex.ru"
host: "api-ps18.herokuapp.com"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /getUserGraphic:
    post:
      tags:
      - "records"
      summary: "Получение график работы пользователя с записями"
      operationId: "getUserGraphic"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID сотрудника"
        required: true
        schema:
          $ref: "#/definitions/body_11"
      responses:
        200:
          description: "Полный график работы, включая записи"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_10_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Records"
    x-swagger-router-controller: "records"
  /setUserGraphic:
    post:
      tags:
      - "records"
      summary: "Установка график работы пользователя"
      operationId: "setUserGraphic"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID сотрудника, массивы удаляемых и/или добавляемых и/или обн\
          овляемых рабочих дней"
        required: true
        schema:
          $ref: "#/definitions/body_12"
      responses:
        200:
          description: "Статус запроса и рузультаты выполнения отдельных операций"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                $ref: "#/definitions/inline_response_200_11_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Records"
    x-swagger-router-controller: "records"
  /setToRecord:
    post:
      tags:
      - "records"
      summary: "Записаться на прием к сотруднику"
      operationId: "setToRecord"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID человека, ID записи, статус Онлайн, способ связи, личный\
          \ контакт"
        required: true
        schema:
          $ref: "#/definitions/body_13"
      responses:
        200:
          description: "Статус запроса и общая информация"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                $ref: "#/definitions/inline_response_200_12_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Records"
    x-swagger-router-controller: "records"
  /cancelRecord:
    post:
      tags:
      - "records"
      summary: "Отмена записи"
      operationId: "cancelRecord"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID записи"
        required: true
        schema:
          $ref: "#/definitions/body_14"
      responses:
        200:
          description: "Статус выполнения запроса"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Records"
    x-swagger-router-controller: "records"
  /skipRecord:
    post:
      tags:
      - "records"
      summary: "Отметить, что клиент не явился по записи"
      operationId: "skipRecord"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID записи"
        required: true
        schema:
          $ref: "#/definitions/body_15"
      responses:
        200:
          description: "Статус выполнения запроса"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Records"
    x-swagger-router-controller: "records"
  /moveRecord:
    post:
      tags:
      - "records"
      summary: "Перенос записи на другое время или день"
      operationId: "moveRecord"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID записи и ID записи на которую переносим"
        required: true
        schema:
          $ref: "#/definitions/body_16"
      responses:
        200:
          description: "Статус запроса и информация о записи"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                $ref: "#/definitions/inline_response_200_12_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Records"
    x-swagger-router-controller: "records"
  /changeRecord:
    post:
      tags:
      - "records"
      summary: "Изменить тип записи и контакты"
      operationId: "changeRecord"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID записи и/или ID онлайн статуса и/или контактные данные"
        required: true
        schema:
          $ref: "#/definitions/body_17"
      responses:
        200:
          description: "Статус выполнения запроса"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Records"
    x-swagger-router-controller: "records"
  /getOneRecord:
    post:
      tags:
      - "records"
      summary: "Получение сведений об одной записи"
      operationId: "getOneRecord"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID записи"
        required: true
        schema:
          $ref: "#/definitions/body_18"
      responses:
        200:
          description: "Статус запроса и данные о записи"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                $ref: "#/definitions/inline_response_200_12_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Records"
    x-swagger-router-controller: "records"
  /getRecordsFromWD:
    post:
      tags:
      - "records"
      summary: "Получить зарегистрированные записи по рабочему дню"
      operationId: "getRecordsFromWD"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID рабочего дня"
        required: true
        schema:
          $ref: "#/definitions/body_19"
      responses:
        200:
          description: "Статус запроса и данные о записях"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_13_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Records"
    x-swagger-router-controller: "records"
  /setJournal:
    post:
      tags:
      - "records"
      summary: "Добавление/Удаление/Изменение записей в журнал посещений"
      operationId: "setJournal"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Массив ID записей в журнале (визитов), которые нужно удалить\
          \ и/или массив визитов, которые нужно добавить, и/или массив визитов, ко\
          торые нужно отредактировать"
        required: true
        schema:
          $ref: "#/definitions/body_20"
      responses:
        200:
          description: "Статус запроса и результаты выполнения отдельных операций"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                $ref: "#/definitions/inline_response_200_14_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Records"
    x-swagger-router-controller: "records"
  /getJournal:
    post:
      tags:
      - "records"
      summary: "Получение записей за период или последние N с шагом F"
      operationId: "getJournal"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Либо получаем записи за период, либо фильтруем по количеству"
        required: true
        schema:
          $ref: "#/definitions/body_21"
      responses:
        200:
          description: "Статус запроса и список записей"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_15_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Records"
    x-swagger-router-controller: "records"
...
  /getRoles:
    post:
      tags:
      - "roles"
      summary: "Получение списка ролей"
      operationId: "getRoles"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Данные созданного пользователя."
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              role_array:
                $ref: "#/definitions/Role"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Roles"
    x-swagger-router-controller: "roles"
  /getUserRoles:
    post:
      tags:
      - "roles"
      summary: "Получить список ролей пользователя"
      operationId: "getUserRoles"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID пользователя"
        required: true
        schema:
          $ref: "#/definitions/body_30"
      responses:
        200:
          description: "Список ролей пользователя"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                $ref: "#/definitions/Role"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Roles"
    x-swagger-router-controller: "roles"
  /addUserRoles:
    post:
      tags:
      - "roles"
      summary: "Добавить роли пользователю"
      operationId: "addUserRoles"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID пользователя и список ID ролей"
        required: true
        schema:
          $ref: "#/definitions/body_31"
      responses:
        200:
          description: "Статус выполнения запроса"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Roles"
    x-swagger-router-controller: "roles"
  /delUserRoles:
    post:
      tags:
      - "roles"
      summary: "Удалить роли у пользователя"
      operationId: "delUserRoles"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID пользователя и список ID ролей"
        required: true
        schema:
          $ref: "#/definitions/body_32"
      responses:
        200:
          description: "Статус выполнения операции."
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Roles"
    x-swagger-router-controller: "roles"
...
securityDefinitions:
  Basic:
    type: "basic"
  JWT:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  Role:
    type: "array"
    items:
      type: "string"
      example: ""
      description: "Value:\n * Admin\n * Student\n * Parent\n * Teacher\n * Psychologist\n"
...
  WorkDay:
    type: "object"
    required:
    - "wd_duration"
    - "wd_time_begin"
    - "wd_time_end"
    properties:
      wd_time_begin:
        type: "string"
        example: "8:00:00"
      wd_time_end:
        type: "string"
        example: "18:30:00"
      wd_break_begin:
        type: "string"
        example: "13:00:00"
      wd_break_end:
        type: "string"
        example: "14:00:00"
      wd_duration:
        type: "integer"
        format: "int32"
        example: 30
    example:
      wd_time_begin: "8:00:00"
      wd_break_begin: "13:00:00"
      wd_break_end: "14:00:00"
      wd_time_end: "18:30:00"
      wd_duration: 30
...
  body_11:
    type: "object"
    required:
    - "emp_id"
    properties:
      emp_id:
        type: "integer"
        format: "int32"
        example: 1
  inline_response_200_10_rec_array:
    required:
    - "rec_data"
    - "rec_id"
    properties:
      rec_id:
        type: "integer"
        format: "int32"
        example: 7
      rec_data:
        $ref: "#/definitions/Record"
    example:
      rec_id: 7
      rec_data:
        rec_time: "11:00:00"
        rec_online: true
        pepl_id: 3
        rec_not_come: false
  inline_response_200_10_payload:
    required:
    - "wd_data"
    - "wd_date"
    - "wd_id"
    properties:
      wd_id:
        type: "integer"
        format: "int32"
        example: 1
      wd_date:
        type: "string"
        format: "date"
        example: "2019-03-03"
      rec_array:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_10_rec_array"
      wd_data:
        $ref: "#/definitions/WorkDay"
    example:
      wd_date: "2019-03-03"
      wd_id: 1
      rec_array:
      - rec_id: 7
        rec_data:
          rec_time: "11:00:00"
          rec_online: true
          pepl_id: 3
          rec_not_come: false
      - rec_id: 7
        rec_data:
          rec_time: "11:00:00"
          rec_online: true
          pepl_id: 3
          rec_not_come: false
      wd_data:
        wd_time_begin: "8:00:00"
        wd_break_begin: "13:00:00"
        wd_break_end: "14:00:00"
        wd_time_end: "18:30:00"
        wd_duration: 30
...
  setUserGraphic_wd_arr_del:
    required:
    - "wd_id"
    properties:
      wd_id:
        type: "integer"
        format: "int32"
        example: 1
  setUserGraphic_wd_arr_add:
    required:
    - "wd_data"
    - "wd_date"
    properties:
      wd_date:
        type: "string"
        format: "date"
        example: "2019-03-03"
      wd_data:
        $ref: "#/definitions/WorkDay"
  setUserGraphic_wd_arr_upd:
    required:
    - "wd_id"
    properties:
      wd_id:
        type: "integer"
        format: "int32"
        example: 1
      wd_time_begin:
        type: "string"
        example: "8:00:00"
      wd_time_end:
        type: "string"
        example: "18:30:00"
      wd_break_begin:
        type: "string"
        example: "13:00:00"
      wd_break_end:
        type: "string"
        example: "14:00:00"
      wd_duration:
        type: "integer"
        format: "int32"
        example: 30
  body_12:
    type: "object"
    required:
    - "emp_id"
    properties:
      emp_id:
        type: "integer"
        format: "int32"
        example: 1
      wd_arr_del:
        type: "array"
        items:
          $ref: "#/definitions/setUserGraphic_wd_arr_del"
      wd_arr_add:
        type: "array"
        items:
          $ref: "#/definitions/setUserGraphic_wd_arr_add"
      wd_arr_upd:
        type: "array"
        items:
          $ref: "#/definitions/setUserGraphic_wd_arr_upd"
  inline_response_200_11_payload_wd_arr_del:
    required:
    - "wd_id"
    - "wd_status"
    properties:
      wd_id:
        type: "integer"
        format: "int32"
        example: 1
      wd_status:
        type: "string"
        example: "OK"
    example:
      wd_id: 1
      wd_status: "OK"
  inline_response_200_11_payload_wd_arr_add:
    required:
    - "wd_date"
    - "wd_status"
    properties:
      wd_date:
        type: "string"
        format: "date"
        example: "2019-03-03"
      wd_status:
        type: "string"
        example: "OK"
    example:
      wd_date: "2019-03-03"
      wd_status: "OK"
  inline_response_200_11_payload:
    properties:
      wd_arr_del:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_11_payload_wd_arr_del"
      wd_arr_add:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_11_payload_wd_arr_add"
      wd_arr_upd:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_11_payload_wd_arr_del"
    example:
      wd_arr_add:
      - wd_date: "2019-03-03"
        wd_status: "OK"
      - wd_date: "2019-03-03"
        wd_status: "OK"
      wd_arr_del:
      - wd_id: 1
        wd_status: "OK"
      - wd_id: 1
        wd_status: "OK"
      wd_arr_upd:
      - wd_id: 1
        wd_status: "OK"
      - wd_id: 1
        wd_status: "OK"
...
  body_13:
    type: "object"
    required:
    - "pepl_id"
    - "rec_id"
    - "rec_online"
    properties:
      pepl_id:
        type: "integer"
        format: "int32"
        example: 3
      rec_id:
        type: "integer"
        format: "int32"
        example: 9
      rec_online:
        type: "boolean"
        example: true
      cont_name:
        type: "string"
        example: "skype"
        description: "Не обязательно,только при личном визите Value:\n  - skype\n\
          \  - discord\n  - hangouts\n  - viber\n  - vk\n  - phone\n"
      cont_value:
        type: "string"
        example: "student132"
  inline_response_200_12_payload:
    required:
    - "emp_id"
    - "pepl_id"
    - "rec_online"
    - "rec_time"
    - "wd_date"
    - "wd_duration"
    properties:
      pepl_id:
        type: "integer"
        format: "int32"
        example: 3
      emp_id:
        type: "integer"
        format: "int32"
        example: 1
      wd_date:
        type: "string"
        format: "date"
        example: "2019-03-03"
      rec_time:
        type: "string"
        example: "12:00:00"
      wd_duration:
        type: "integer"
        format: "int32"
        example: 30
      rec_online:
        type: "boolean"
        example: true
      cont_name:
        type: "string"
        example: "skype"
        description: "Не обязательно,только при личном визите Value:\n  - skype\n\
          \  - discord\n  - hangouts\n  - viber\n  - vk\n  - phone\n"
      cont_value:
        type: "string"
        example: "student132"
    example:
      rec_time: "12:00:00"
      rec_online: true
      wd_date: "2019-03-03"
      pepl_id: 3
      cont_value: "student132"
      wd_duration: 30
      emp_id: 1
      cont_name: "skype"
...
  body_14:
    type: "object"
    required:
    - "rec_id"
    properties:
      rec_id:
        type: "integer"
        format: "int32"
        example: 9
  body_15:
    type: "object"
    required:
    - "rec_id"
    properties:
      rec_id:
        type: "integer"
        format: "int32"
        example: 9
  body_16:
    type: "object"
    required:
    - "new_rec_id"
    - "rec_id"
    properties:
      rec_id:
        type: "integer"
        format: "int32"
        example: 9
      new_rec_id:
        type: "integer"
        format: "int32"
        example: 10
  body_17:
    type: "object"
    required:
    - "rec_id"
    properties:
      rec_id:
        type: "integer"
        format: "int32"
        example: 9
      rec_online:
        type: "boolean"
        example: false
      cont_name:
        type: "string"
        example: "phone"
        description: "Value:\n  - skype\n  - discord\n  - hangouts\n  - viber\n  -\
          \ vk\n  - phone\n"
      cont_value:
        type: "string"
        example: "9652354488"
  body_18:
    type: "object"
    required:
    - "rec_id"
    properties:
      rec_id:
        type: "integer"
        format: "int32"
        example: 9
  body_19:
    type: "object"
    required:
    - "wd_id"
    properties:
      wd_id:
        type: "integer"
        format: "int32"
        example: 1
  inline_response_200_13_payload:
    required:
    - "emp_id"
    - "pepl_id"
    - "rec_id"
    - "rec_online"
    - "rec_time"
    - "wd_date"
    - "wd_duration"
    properties:
      rec_id:
        type: "integer"
        format: "int32"
        example: 9
      pepl_id:
        type: "integer"
        format: "int32"
        example: 3
      emp_id:
        type: "integer"
        format: "int32"
        example: 1
      wd_date:
        type: "string"
        format: "date"
        example: "2019-03-03"
      rec_time:
        type: "string"
        example: "12:00:00"
      wd_duration:
        type: "integer"
        format: "int32"
        example: 30
      rec_online:
        type: "boolean"
        example: true
      cont_name:
        type: "string"
        example: "skype"
        description: "Не обязательно,только при личном визите Value:\n  - skype\n\
          \  - discord\n  - hangouts\n  - viber\n  - vk\n  - phone\n"
      cont_value:
        type: "string"
        example: "student132"
    example:
      rec_time: "12:00:00"
      rec_online: true
      wd_date: "2019-03-03"
      pepl_id: 3
      cont_value: "student132"
      rec_id: 9
      wd_duration: 30
      emp_id: 1
      cont_name: "skype"
...
  setJournal_vst_arr_del:
    required:
    - "vst_id"
    properties:
      vst_id:
        type: "integer"
        format: "int32"
        example: 1
  setJournal_rec_data:
    required:
    - "rec_id"
    properties:
      rec_id:
        type: "integer"
        format: "int32"
        example: 7
    description: "либо rec_data, либо vst_data"
...
  setJournal_vst_arr_add:
    required:
    - "vst_problem"
    - "vst_reason"
    - "vst_result"
    properties:
      vst_reason:
        type: "string"
        example: "Стресс"
      vst_problem:
        type: "string"
        example: "Угнетение учениками"
      vst_result:
        type: "string"
        example: "Стресс"
      rec_data:
        $ref: "#/definitions/setJournal_rec_data"
      vst_data:
        $ref: "#/definitions/setJournal_vst_data"
  setJournal_rec_data_1:
    properties:
      rec_id:
        type: "integer"
        format: "int32"
        example: 7
    description: "либо rec_data, либо vst_data"
...
  setJournal_vst_arr_upd:
    required:
    - "vst_id"
    properties:
      vst_id:
        type: "integer"
        format: "int32"
        example: 1
      vst_reason:
        type: "string"
        example: "Стресс"
      vst_problem:
        type: "string"
        example: "Угнетение учениками"
      vst_result:
        type: "string"
        example: "Стресс"
      rec_data:
        $ref: "#/definitions/setJournal_rec_data_1"
      vst_data:
        $ref: "#/definitions/setJournal_vst_data_1"
  body_20:
    type: "object"
    properties:
      vst_arr_del:
        type: "array"
        items:
          $ref: "#/definitions/setJournal_vst_arr_del"
      vst_arr_add:
        type: "array"
        items:
          $ref: "#/definitions/setJournal_vst_arr_add"
      vst_arr_upd:
        type: "array"
        items:
          $ref: "#/definitions/setJournal_vst_arr_upd"
  inline_response_200_14_payload_vst_arr_del:
    required:
    - "vst_id"
    - "vst_status"
    properties:
      vst_id:
        type: "integer"
        format: "int32"
        example: 1
      vst_status:
        type: "string"
        example: "OK"
    example:
      vst_status: "OK"
      vst_id: 1
...
  inline_response_200_14_payload_vst_arr_add:
    required:
    - "vst_status"
    properties:
      vst_status:
        type: "string"
        example: "OK"
      rec_data:
        $ref: "#/definitions/inline_response_200_14_payload_rec_data"
      vst_data:
        $ref: "#/definitions/inline_response_200_14_payload_vst_data"
    example:
      vst_data:
        vst_dt: "2019-03-03T11:00:00.000Z"
      vst_status: "OK"
      rec_data:
        rec_id: 7
  inline_response_200_14_payload:
    properties:
      vst_arr_del:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_14_payload_vst_arr_del"
      vst_arr_add:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_14_payload_vst_arr_add"
      vst_arr_upd:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_14_payload_vst_arr_del"
    example:
      vst_arr_upd:
      - vst_status: "OK"
        vst_id: 1
      - vst_status: "OK"
        vst_id: 1
      vst_arr_del:
      - vst_status: "OK"
        vst_id: 1
      - vst_status: "OK"
        vst_id: 1
      vst_arr_add:
      - vst_data:
          vst_dt: "2019-03-03T11:00:00.000Z"
        vst_status: "OK"
        rec_data:
          rec_id: 7
      - vst_data:
          vst_dt: "2019-03-03T11:00:00.000Z"
        vst_status: "OK"
        rec_data:
          rec_id: 7
...
  getJournal_byDate:
    required:
    - "vst_date_begin"
    properties:
      vst_date_begin:
        type: "string"
        format: "date"
        example: "2019-02-01"
      vst_date_end:
        type: "string"
        format: "date"
        example: "2019-02-28"
  getJournal_inCount:
    required:
    - "vst_N"
    properties:
      vst_N:
        type: "integer"
        format: "int32"
        example: 50
      vst_F:
        type: "integer"
        format: "int32"
        example: 0
    description: "Получить N записей с шагом F"
  body_21:
    type: "object"
    properties:
      byDate:
        $ref: "#/definitions/getJournal_byDate"
      inCount:
        $ref: "#/definitions/getJournal_inCount"
  inline_response_200_15_payload:
    required:
    - "vst_age"
    - "vst_consultant"
    - "vst_dt"
    - "vst_gender"
    - "vst_id"
    - "vst_name"
    - "vst_problem"
    - "vst_reason"
    - "vst_result"
    properties:
      vst_id:
        type: "integer"
        format: "int32"
        example: 1
      rec_id:
        type: "integer"
        format: "int32"
        example: 7
      vst_dt:
        type: "string"
        format: "date-time"
        example: "2019-03-03T11:00:00.000Z"
      vst_age:
        type: "integer"
        format: "int32"
        example: 13
      vst_gender:
        type: "boolean"
        example: true
      vst_name:
        type: "string"
        example: "Тарасов Семен Львович"
      vst_reason:
        type: "string"
        example: "Стресс"
      vst_problem:
        type: "string"
        example: "Угнетение учениками"
      vst_result:
        type: "string"
        example: "Стресс"
      vst_consultant:
        type: "string"
        example: "Иванов Иван иванович"
    example:
      vst_problem: "Угнетение учениками"
      vst_name: "Тарасов Семен Львович"
      vst_id: 1
      vst_consultant: "Иванов Иван иванович"
      vst_reason: "Стресс"
      vst_result: "Стресс"
      rec_id: 7
      vst_age: 13
      vst_dt: "2019-03-03T11:00:00.000Z"
      vst_gender: true
...
  body_30:
    type: "object"
    required:
    - "pepl_id"
    properties:
      pepl_id:
        type: "integer"
        format: "int32"
        example: 1
...
  body_31:
    type: "object"
    required:
    - "pepl_id"
    - "role_array"
    properties:
      pepl_id:
        type: "integer"
        format: "int32"
        example: 1
      role_array:
        $ref: "#/definitions/Role"
  body_32:
    type: "object"
    required:
    - "pepl_id"
    - "role_array"
    properties:
      pepl_id:
        type: "integer"
        format: "int32"
        example: 1
      role_array:
        $ref: "#/definitions/Role"
...